fix: renamed folder app to components
